Apply Clean (Onion) Architecture Remove using of repository pattern Add service registration per layer
