+++ Application/DependencyInjection.cs
@@ -1,0 +1,11 @@
+﻿namespace UserListsAPI.Application;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddApplicationServices(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddAutoMapper(typeof(AppMappingProfile));
+
+        return services;
+    }
+}
