+++ Application/Exceptions/AppException.cs
@@ -1,0 +1,12 @@
+﻿using System.Globalization;
+
+namespace UserListsAPI.Application.Exceptions;
+
+public class AppException : Exception
+{
+    public AppException() : base() { }
+
+    public AppException(string message) : base(message) { }
+
+    public AppException(string message, params object[] args) : base(string.Format(CultureInfo.CurrentCulture, message, args)) { }
+}
