--- Data/Enums/ItemStatus.cs
@@ -1,8 +1,0 @@
-﻿namespace UserListsAPI.Data.Enums;
-
-public enum ItemStatus
-{
-    Ok,
-    Unavailable,
-    Irrelevant
-}
