+++ DependencyInjection.cs
@@ -1,0 +1,61 @@
+﻿using Microsoft.AspNetCore.Mvc.Versioning;
+using Microsoft.OpenApi.Models;
+
+namespace UserListsAPI;
+
+public static class DependencyInjection
+{
+    public static IServiceCollection AddPresentationServices(this IServiceCollection services, IConfiguration configuration)
+    {
+        string host = configuration.GetValue<string>("UserListsMVC:Host");
+        int port = configuration.GetValue<int>("UserListsMVC:Port");
+        services.AddCors(options => options.AddDefaultPolicy(policy => policy.WithOrigins($"https://{host}:{port}").AllowAnyMethod().AllowAnyHeader()));
+
+        services.AddControllers();
+
+        services.AddEndpointsApiExplorer();
+
+        services.AddSwaggerGen(options =>
+        {
+            options.AddSecurityDefinition("ApiKey", new OpenApiSecurityScheme
+            {
+                Description = "ApiKey must appear in header",
+                Type = SecuritySchemeType.ApiKey,
+                Name = "XApiKey",
+                In = ParameterLocation.Header,
+                Scheme = "ApiKeyScheme"
+            });
+            var key = new OpenApiSecurityScheme()
+            {
+                Reference = new OpenApiReference
+                {
+                    Type = ReferenceType.SecurityScheme,
+                    Id = "ApiKey"
+                },
+                In = ParameterLocation.Header
+            };
+            var requirement = new OpenApiSecurityRequirement { { key, new List<string>() } };
+            options.AddSecurityRequirement(requirement);
+        });
+
+        services.AddApiVersioning(options =>
+        {
+            options.AssumeDefaultVersionWhenUnspecified = true;
+            options.DefaultApiVersion = new Microsoft.AspNetCore.Mvc.ApiVersion(1, 0);
+            options.ReportApiVersions = true;
+            options.ApiVersionReader = ApiVersionReader.Combine(
+          new UrlSegmentApiVersionReader(),
+          new HeaderApiVersionReader("XApiVersion"));
+        });
+
+        services.AddVersionedApiExplorer(setup =>
+        {
+            setup.GroupNameFormat = "'v'VVV";
+            setup.SubstituteApiVersionInUrl = true;
+        });
+
+        services.ConfigureOptions<ConfigureSwaggerOptions>();
+
+        return services;
+    }
+}
