+++ Domain/Enums/ItemStatus.cs
@@ -1,0 +1,8 @@
+﻿namespace UserListsAPI.Domain.Enums;
+
+public enum ItemStatus
+{
+    Ok,
+    Unavailable,
+    Irrelevant
+}
