+++ Infrastructure/Api/JsonModels/GameReviewsJson.cs
@@ -1,0 +1,15 @@
+﻿using System.Text.Json.Serialization;
+
+namespace UserListsAPI.Infrastructure.Api.JsonModels;
+
+public record GameReviewsJson
+{
+    [JsonPropertyName("review_score_desc")]
+    public string ReviewScore { get; set; } = null!;
+    [JsonPropertyName("total_positive")]
+    public int TotalPositive { get; set; }
+    [JsonPropertyName("total_negative")]
+    public int TotalNegative { get; set; }
+    [JsonPropertyName("total_reviews")]
+    public int TotalReviews { get; set; }
+}
