--- JsonModels/GameJsonShort.cs
@@ -1,13 +1,0 @@
-﻿using System.Text.Json.Serialization;
-
-namespace UserListsAPI.JsonModels;
-
-public record GameJsonShort
-{
-    [JsonPropertyName("appid")]
-    [JsonConverter(typeof(StringJsonConverter))]
-    public string Id { get; set; } = null!;
-
-    [JsonPropertyName("name")]
-    public string Title { get; set; } = null!;
-}
